requirements.txt
# Core dependencies
numpy>=1.20.0
h5py>=3.1.0
faiss-cpu>=1.7.2  # Use faiss-gpu for GPU support
torch>=1.9.0
tensorly>=0.6.0
flask>=2.0.1
flask-cors>=3.0.10
gunicorn>=20.1.0

# API and utilities
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
typing-extensions>=3.10.0

# Testing and development
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.6b0
flake8>=3.9.2

# Visualization 
matplotlib>=3.4.2
seaborn>=0.11.1

# Optional integrations
tensorflow>=2.6.0  # Optional for TensorFlow integration
scikit-learn>=0.24.2  # For dimensionality reduction 